Add time module to API
