+++ src/api/mod.rs
@@ -45,5 +45,6 @@ pub mod prompt;
 pub mod random;
 pub mod regex;
 pub mod syscall;
+pub mod time;
 pub mod vga;
 // TODO: add mod wildcard
